src/main/java/com/digitalacademy/customer/service/CustomerService.java
package com.digitalacademy.customer.service;

import com.digitalacademy.customer.model.Customer;
import com.digitalacademy.customer.repositories.CustomerRepository;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.dao.EmptyResultDataAccessException;
import org.springframework.stereotype.Service;

import java.util.List;

@Service
public class CustomerService {

    private CustomerRepository customerRepository;

    @Autowired
    public CustomerService(CustomerRepository customerRepository){
        this.customerRepository = customerRepository;
    }
    public List<Customer> getCustomerList(){
        return customerRepository.findAll();
    }
    public Customer getCustomerById(Long id){
        return customerRepository.findAllById(id);
    }
    public List<Customer> getCustomerByFirstName(String firstName){
        return customerRepository.findByFirstName(firstName);
    }
    public Customer createCustomer(Customer customer){
        return customerRepository.save(customer);
    }
    public Customer updateCustomer(Long id, Customer customerReq){
        return customerRepository.findAllById(id) != null ?
                customerRepository.save(customerReq) : null;
    }
    public boolean deleteById(Long id){
        try{
            customerRepository.deleteById(id);
            return true;
        }
        catch (EmptyResultDataAccessException e){
            return false;
        }
    }
}
